Cooldown starter moved to app start.
